Finished moving photaatomic files around
