+++ cmake/unit_testing.cmake
@@ -84,9 +84,10 @@ add_subdirectory( src/ACEtk/photoatomic/PrincipalCrossSectionBlock/test )
 add_subdirectory( src/ACEtk/photoatomic/IncoherentScatteringFunctionBlock/test )
 add_subdirectory( src/ACEtk/photoatomic/CoherentFormFactorBlock/test )
 add_subdirectory( src/ACEtk/photoatomic/FluorescenceDataBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoatomicHeatingNumbersBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoatomicComptonProfile/test )
-add_subdirectory( src/ACEtk/block/PhotoatomicComptonProfileBlock/test )
+add_subdirectory( src/ACEtk/photoatomic/HeatingNumbersBlock/test )
+add_subdirectory( src/ACEtk/photoatomic/ComptonProfile/test )
+add_subdirectory( src/ACEtk/photoatomic/ComptonProfileBlock/test )
+add_subdirectory( src/ACEtk/photoatomic/PhotoelectricCrossSectionBlock/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicElectronShellBlock/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicElectronSubshellBlock/test )
 add_subdirectory( src/ACEtk/block/PhotonuclearPrincipalCrossSectionBlock/test )
@@ -97,7 +98,6 @@ add_subdirectory( src/ACEtk/block/PhotoatomicElectronExcitationBlock/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicElectronBremsstrahlungBlock/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicSubshellTransitionData/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicSubshellTransitionDataBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoelectricCrossSectionBlock/test )
 add_subdirectory( src/ACEtk/block/ElectronTabulatedAngularDistribution/test )
 add_subdirectory( src/ACEtk/block/ElectronElasticAngularDistributionBlock/test )
 add_subdirectory( src/ACEtk/block/ElectronTabulatedEnergyDistribution/test )
