+++ python/src/photoatomic/ComptonProfileBlock.python.cpp
@@ -3,7 +3,7 @@
 #include <pybind11/stl.h>
 
 // local includes
-#include "ACEtk/block/PhotoatomicComptonProfileBlock.hpp"
+#include "ACEtk/photoatomic/ComptonProfileBlock.hpp"
 #include "views.hpp"
 #include "definitions.hpp"
