+++ python/test/CMakeLists.txt
@@ -66,16 +66,16 @@ add_test( NAME ACEtk.python.photoatomic.PrincipalCrossSectionBlock COMMAND ${PYT
 add_test( NAME ACEtk.python.photoatomic.IncoherentScatteringFunctionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_IncoherentScatteringFunctionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.photoatomic.CoherentFormFactorBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_CoherentFormFactorBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.photoatomic.FluorescenceDataBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_PhotoatomicFluorescenceDataBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.PhotoatomicHeatingNumbersBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicHeatingNumbersBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.PhotoatomicComptonProfile COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicComptonProfile.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.PhotoatomicComptonProfileBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicComptonProfileBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.HeatingNumbersBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_PhotoatomicHeatingNumbersBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.PhotoatomicComptonProfile COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_PhotoatomicComptonProfile.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.PhotoatomicComptonProfileBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_PhotoatomicComptonProfileBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.PhotoelectricCrossSectionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_PhotoelectricCrossSectionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicElectronShellBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicElectronShellBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicElectronSubshellBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicElectronSubshellBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicElectronCrossSectionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicElectronCrossSectionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicElectronElasticCrossSectionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicElectronElasticCrossSectionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicElectronExcitationBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicElectronExcitationBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicElectronBremsstrahlungBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicElectronExcitationBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.PhotoelectricCrossSectionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoelectricCrossSectionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotonuclearPrincipalCrossSectionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotonuclearPrincipalCrossSectionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotonuclearSecondaryParticleLocatorBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotonuclearSecondaryParticleLocatorBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicSubshellTransitionData COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicSubshellTransitionData.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
