+++ src/ACEtk/PhotoatomicTable.hpp
@@ -9,11 +9,12 @@
 #include "ACEtk/photoatomic/IncoherentScatteringFunctionBlock.hpp"
 #include "ACEtk/photoatomic/CoherentFormFactorBlock.hpp"
 #include "ACEtk/photoatomic/FluorescenceDataBlock.hpp"
-#include "ACEtk/block/PhotoatomicHeatingNumbersBlock.hpp"
+#include "ACEtk/photoatomic/HeatingNumbersBlock.hpp"
+#include "ACEtk/photoatomic/ComptonProfileBlock.hpp"
+#include "ACEtk/photoatomic/PhotoelectricCrossSectionBlock.hpp"
+
 #include "ACEtk/block/PhotoatomicElectronShellBlock.hpp"
-#include "ACEtk/block/PhotoatomicComptonProfileBlock.hpp"
 #include "ACEtk/block/PhotoatomicElectronSubshellBlock.hpp"
-#include "ACEtk/block/PhotoelectricCrossSectionBlock.hpp"
 #include "ACEtk/block/PhotoatomicSubshellTransitionDataBlock.hpp"
 #include "ACEtk/block/PhotoatomicElectronCrossSectionBlock.hpp"
 #include "ACEtk/block/PhotoatomicElectronElasticCrossSectionBlock.hpp"
