+++ src/ACEtk/photoatomic/ComptonProfile/test/PhotoatomicComptonProfile.test.cpp
@@ -1,7 +1,7 @@
 #define CATCH_CONFIG_MAIN
 
 #include "catch.hpp"
-#include "ACEtk/block/PhotoatomicComptonProfile.hpp"
+#include "ACEtk/photoatomic/ComptonProfile.hpp"
 
 // other includes
