+++ src/ACEtk/photoatomic/PhotoelectricCrossSectionBlock/test/PhotoelectricCrossSectionBlock.test.cpp
@@ -1,7 +1,7 @@
 #define CATCH_CONFIG_MAIN
 
 #include "catch.hpp"
-#include "ACEtk/block/PhotoelectricCrossSectionBlock.hpp"
+#include "ACEtk/photoatomic/PhotoelectricCrossSectionBlock.hpp"
 
 // other includes
